requirements.txt
accelerate==1.11.0
bitsandbytes==0.48.2
datasets==3.6.0
diffusers==0.35.2
fairscale==0.4.13
gradio==6.0.1
huggingface_hub==0.36.0
modules==1.0.0
numpy==1.26.4
packaging==25.0
peft==0.17.1
Pillow==12.0.0
pycocoevalcap==1.2
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
tensorboardX==2.6.4
timm==1.0.22
torch==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.57.1
wandb==0.23.0


git+https://github.com/openai/CLIP.git
git+https://github.com/jmhessel/pycocoevalcap.git
open-clip-torch
clip @ git+https://github.com/openai/CLIP.git
